installer/pkg: Move terraform var definition to pkg/typs/config.
