+++ installer/pkg/config-generator/BUILD.bazel
@@ -10,12 +10,12 @@ go_library(
     importpath = "github.com/openshift/installer/installer/pkg/config-generator",
     visibility = ["//visibility:public"],
     deps = [
-        "//installer/pkg/config:go_default_library",
         "//installer/pkg/copy:go_default_library",
         "//pkg/asset/tls:go_default_library",
         "//pkg/ipnet:go_default_library",
         "//pkg/rhcos:go_default_library",
         "//pkg/types:go_default_library",
+        "//pkg/types/config:go_default_library",
         "//vendor/github.com/apparentlymart/go-cidr/cidr:go_default_library",
         "//vendor/github.com/coreos/ignition/config/v2_2:go_default_library",
         "//vendor/github.com/coreos/ignition/config/v2_2/types:go_default_library",
@@ -36,8 +36,8 @@ go_test(
     data = glob(["fixtures/**"]),
     embed = [":go_default_library"],
     deps = [
-        "//installer/pkg/config:go_default_library",
         "//pkg/asset/tls:go_default_library",
+        "//pkg/types/config:go_default_library",
         "//vendor/github.com/stretchr/testify/assert:go_default_library",
     ],
 )
