+++ installer/pkg/workflow/BUILD.bazel
@@ -15,9 +15,9 @@ go_library(
     importpath = "github.com/openshift/installer/installer/pkg/workflow",
     visibility = ["//visibility:public"],
     deps = [
-        "//installer/pkg/config:go_default_library",
         "//installer/pkg/config-generator:go_default_library",
         "//installer/pkg/copy:go_default_library",
+        "//pkg/types/config:go_default_library",
         "//vendor/github.com/Sirupsen/logrus:go_default_library",
         "//vendor/gopkg.in/yaml.v2:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
@@ -37,5 +37,5 @@ go_test(
     ],
     data = glob(["fixtures/**"]),
     embed = [":go_default_library"],
-    deps = ["//installer/pkg/config:go_default_library"],
+    deps = ["//pkg/types/config:go_default_library"],
 )
