+++ pkg/rhcos/ami.go
@@ -4,6 +4,11 @@ import (
 	"fmt"
 )
 
+const (
+	// DefaultChannel is the default RHCOS channel for the cluster.
+	DefaultChannel = "tested"
+)
+
 // AMI calculates a Red Hat CoreOS AMI.
 func AMI(channel, region string) (ami string, err error) {
 	if channel != "tested" {
