+++ pkg/types/config/aws/BUILD.bazel
@@ -3,6 +3,6 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["aws.go"],
-    importpath = "github.com/openshift/installer/installer/pkg/config/aws",
+    importpath = "github.com/openshift/installer/pkg/types/config/aws",
     visibility = ["//visibility:public"],
 )
